refactored common settings
