+++ JeMPI_Apps/JeMPI_UI/src/pages/settings/common/Common.tsx
@@ -9,22 +9,13 @@ import {
   GridRowModes,
   GridRowModesModel,
   GridActionsCellItem,
-  GridRowsProp,
-  GridToolbarContainer
 } from '@mui/x-data-grid'
 import EditIcon from '@mui/icons-material/Edit'
 import SaveIcon from '@mui/icons-material/Save'
 import CancelIcon from '@mui/icons-material/Close'
-import { useSnackbar } from 'notistack'
 import { useEffect, useState } from 'react'
-import { randomId } from 'utils/helpers'
+import { EditToolbar } from 'components/shared/EditToolBar'
 
-interface EditToolbarProps {
-  setRows: (newRows: (oldRows: GridRowsProp) => GridRowsProp) => void
-  setRowModesModel: (
-    newModel: (oldModel: GridRowModesModel) => GridRowModesModel
-  ) => void
-}
 
 const CommonSettings = ({ demographicData }: { demographicData: any }) => {
 
@@ -229,21 +220,6 @@ const CommonSettings = ({ demographicData }: { demographicData: any }) => {
       )}
     </Box>
   )
-}
-
-function EditToolbar(props: EditToolbarProps) {
-  const { setRows, setRowModesModel } = props
-
-  const handleClick = () => {
-    const id = randomId()
-    setRows(oldRows => [...oldRows, { id, name: '', age: '', isNew: true }])
-    setRowModesModel(oldModel => ({
-      ...oldModel,
-      [id]: { mode: GridRowModes.Edit, fieldToFocus: 'name' }
-    }))
-  }
-
-  return <GridToolbarContainer></GridToolbarContainer>
 }
 
 export default CommonSettings
